Refactored welcome to use data-binding
